# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/order": {
            "post": {
                "description": "Creates a new order in the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Place Order",
                "parameters": [
                    {
                        "description": "Order Data",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.Order"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "description": "Retrieves a list of products from the service.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get Products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Product"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entities.Customer": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "г. Москва, ул. Ленина, д. 1"
                },
                "name": {
                    "type": "string",
                    "example": "Иван Иванов"
                },
                "phone": {
                    "type": "string",
                    "example": "+79991234567"
                }
            }
        },
        "entities.Manager": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "mail@gmail.com"
                },
                "phone": {
                    "type": "string",
                    "example": "+79991234567"
                }
            }
        },
        "entities.Order": {
            "type": "object",
            "properties": {
                "customer": {
                    "$ref": "#/definitions/entities.Customer"
                },
                "date": {
                    "type": "integer",
                    "example": 1630000000
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "manager": {
                    "$ref": "#/definitions/entities.Manager"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.ProductInOrder"
                    }
                }
            }
        },
        "entities.Product": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "img": {
                    "type": "string",
                    "example": "https://example.com/img.jpg"
                },
                "lifetime": {
                    "type": "integer",
                    "example": 86400
                },
                "name": {
                    "type": "string",
                    "example": "Эклеры \"Домашние\" (с домашним кремом) (ЛОРИ)"
                },
                "weight": {
                    "type": "integer",
                    "example": 1000
                }
            }
        },
        "entities.ProductInOrder": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 2
                },
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    }
}